Add a way to configure how queues are declared to handle complex scenario.
It's now possible to pass an implementation of `IRabbitMQQueueDefinitionFactory`
to a trigger so when the queue is declared, it will use the definition returned
by the methods.

This allow to configure Durable Queue (fixes: #68)
This may help with #76 since it allow to fully configure queue

This allow scenario like in the new Sample: where we may want to have 3 queue
chained together to handle error nicely, when the first queue fail to process
a message, it goes to the second queue, this allow to handle error after all
retry. And if this failed again, it can go to another queue that may be handle
manually and that will contains unexpetected errors.


Queue => Queue-Failed => Queue-Error